Package manual

Package manual

Types

Model types

Results

Functions

Model helpers

Simulation

Bayesian inference

Bayesian model analysis

Utilities

Visualisation

Advanced features.

Index

Types

Model types

Missing docstring.

Missing docstring for DPOMPModel. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Particle. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Event. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Observation. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ARQModel. Check Documenter's build log for details.

Results

Missing docstring.

Missing docstring for SimResults. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ImportanceSample. Check Documenter's build log for details.

Missing docstring.

Missing docstring for RejectionSample. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MCMCSample. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ARQMCMCSample. Check Documenter's build log for details.

Functions

Model helpers

BayesianWorkflows.generate_model — Function

generate_model(model_name, initial_condition; freq_dep = false, obs_error = 2.0)

Generates an DPOMPModel instance. Observation models are generated using the partial_gaussian_obs_model function, with `σ = obs_error (see that functions entry for further details.)

Parameters

model_name – the model, e.g. "SI"; "SIR"; "SEIR"; etc

initial_condition – initial condition.

Named parameters

freq_dep – epidemiological models only, set to true for frequency-dependent contact rates.

obs_error – average observation error (default = 2.)

t0_index – index of the parameter that represents the initial time. 0 if fixed at 0.0.

model_name options

"SI"

"SIR"

"SIS"

"SEI"

"SEIR"

"SEIS"

"SEIRS"

"PREDPREY"

"ROSSMAC"

Examples

generate_model("SIS", [100,1])

Missing docstring.

Missing docstring for generate_custom_model. Check Documenter's build log for details.

Missing docstring.

Missing docstring for partial_gaussian_obs_model. Check Documenter's build log for details.

Simulation

BayesianWorkflows.gillespie_sim — Function

gillespie_sim(model, parameters; tmax = 100.0, num_obs = 5)

Run a Doob-Gillespie (DGA) simulation based on model.

Returns a SimResults type containing the trajectory and observations data, or an array of the same if n_sims > 1.

Parameters

model – DPOMPModel (see [DCTMPs.jl models]@ref).

parameters – model parameters.

Optional

tmax – maximum time (default: 100.)

n_obs – the number of observations to draw (default: 5.)

n_sims – number of simulations to draw (default: 1.)

Example

using DiscretePOMP
m = generate_model("SIR", [50, 1, 0])
x = DiscretePOMP.gillespie_sim(model, [0.005, 0.12])
println(DiscretePOMP.plot_trajectory(x))

Bayesian inference

Missing docstring.

Missing docstring for run_mcmc_analysis. Check Documenter's build log for details.

Missing docstring.

Missing docstring for run_ibis_analysis. Check Documenter's build log for details.

Missing docstring.

Missing docstring for run_arq_mcmc_analysis. Check Documenter's build log for details.

Bayesian model analysis

Missing docstring.

Missing docstring for run_model_comparison_analysis. Check Documenter's build log for details.

Utilities

BayesianWorkflows.get_observations — Function

get_observations(source)

Return an array of type Observation, based on a two-dimensional array, DataFrame or file location (i.e. String.)

Note that a observation times must be in the first column of the input variable.

BayesianWorkflows.tabulate_results — Function

tabulate_results(results)

Display the results of an inference analysis.

The main parameter is results – a data structure of type MCMCSample, ImportanceSample, ARQMCMCSample or ModelComparisonResults. When invoking the latter, the named parameter null_index = 1 by default but can be overridden. This determines the 'null' model, used to compute the Bayes factor.

Missing docstring.

Missing docstring for save_to_file. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_particle_filter_lpdf. Check Documenter's build log for details.

Visualisation

BayesianWorkflows.plot_trajectory — Function

plot_trajectory(x; plot_index=[:])

Plot the trajectory of a a DGA simulation using UnicodePlots.jl.

The only input parameter required is x of type SimResults, i.e. from a call to gillespie_sim. All system states are plotted by default, but a subset can be specified by passing an integer array to the plot_index option, which contains the indices of the desired subset. E.g. [1,2] for the first two 'compartments' only.

BayesianWorkflows.plot_parameter_trace — Function

plot_parameter_trace(mcmc, [parameter::Int64])

Produce a trace plot of samples using UnicodePlots.jl.

The mcmc input is of type MCMCSample, ARQMCMCSample or RejectionSample. The parameter index can be optionally specified, else all parameters are plotted and returned as an Array of unicode plots.

Missing docstring.

Missing docstring for plot_parameter_marginal. Check Documenter's build log for details.

BayesianWorkflows.plot_parameter_heatmap — Function

plot_parameter_heatmap(mcmc, x_parameter, y_parameter)

Plot the marginal distribution of samples from an MCMC analysis for two model parameters using UnicodePlots.jl.

Parameters

mcmc – MCMCResults, e.g. from a call to run_met_hastings_mcmc.

x_parameter – the index of the model parameter to be plotted on the x axis.

y_parameter – the index of the model parameter to be plotted on the y axis.

BayesianWorkflows.plot_model_comparison — Function

plot_model_comparison(results; boxplot = true)

Plot the Bayesian model evidence (BME) from a model comparison analysis, using UnicodePlots.jl.

Parameters

results – ModelComparisonResults, i.e. from a call to run_model_comparison_analysis.

boxplot – true for a series of boxplots, else a simple UnicodePlots.barplot showing only the average BME for each model variant (default.)

Advanced features.

This section covers functionality for customising the algorithms themselves.

Missing docstring.

Missing docstring for run_custom_mcmc_analysis. Check Documenter's build log for details.

Missing docstring.

Missing docstring for generate_custom_particle. Check Documenter's build log for details.

Index

BayesianWorkflows.generate_model

BayesianWorkflows.get_observations

BayesianWorkflows.gillespie_sim

BayesianWorkflows.plot_model_comparison

BayesianWorkflows.plot_parameter_heatmap

BayesianWorkflows.plot_parameter_trace

BayesianWorkflows.plot_trajectory

BayesianWorkflows.tabulate_results