DPOMP models

Discrete-state space Partially Observed Markov Processes

Predefined models

The package includes the following (mostly) epidemiological models as predefined examples:

"SI"

"SIR"

"SIS"

"SEI"

"SEIR"

"SEIS"

"SEIRS"

"PREDPREY"

"ROSSMAC"

They can be instantiated using the generate_model function:

using BayesianWorkflows
model_name = "SIS"
initial_model_state = [100, 1]
model = generate_model(model_name, initial_model_state)

NB. this generates a standard configuration of the named model that can be tweaked later.

Options

Aside from the model name and initial model state (i.e. the initial 'population',) there are a number of options for generating a default model configuration:

freq_dep – epidemiological models only, set to true for frequency-dependent contact rates.

obs_error – average observation error (default = 2.)

t0_index – index of the parameter that represents the initial time. 0 if fixed at 0.0.

Full model configuration

BayesianWorkflows.DPOMPModel — Type

DPOMPModel

A mutable struct which represents a DSSCT model (see Models for further details).

Fields

name – string, e,g, "SIR".

n_events – number of distinct event types.

event_rates! – event rate function.

initial_state – function for sampling initial model (i.e. population) state.

transition! – transition function of form f!(population, event_type).

obs_model – observation model likelihood function.

obs_function – observation function, use this to add 'noise' to simulated observations.

t0_index – index of the parameter that represents the initial time. 0 if fixed at 0.0.